requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.27
django-allauth==65.13.1
django-crispy-forms==2.5
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
whitenoise==6.11.0
razorpay==2.0.0
django-anymail==14.0
sentry-sdk==2.19.2
markdown==3.9
bleach==6.1.0
django-recaptcha==4.0.0
django-webpush==0.3.6
coverage
